Ldt xx fix game selection (#50)
* LDT-xx: quick fix for library games selection of not-checked-in games

* LDT-xx: Added dashboard page. Added popup to show full image of game.

+++ backend/api/models.py
@@ -133,7 +133,7 @@ class StoreGame(models.Model):
     )
     selling_price = models.FloatField(default=0.0, blank=False, null=False)
     buying_price = models.FloatField(default=0.0, blank=False, null=False)
-    stock = models.FloatField(default=0.0, blank=False, null=False)
+    stock = models.IntegerField(default=0, blank=False, null=False)
 
 
 class Withdraw(models.Model):
